==> 1. ETL.ipynb <==
import pandas as pd 
import ast

def cargar_json(ruta_archivo):
    """
    Se carga el archivo JSON y se devuelve un DataFrame.

    Parámetros:
        ruta_archivo (str): La ruta al archivo JSON.

    Retorno:
        pd.DataFrame: El DataFrame cargado desde el archivo JSON.
    """
    filas = []

    # Abrir el archivo en modo lectura con la codificación "MacRoman"
    with open(ruta_archivo, encoding="MacRoman") as archivo:
        # Iterar sobre cada línea en el archivo
        for linea in archivo.readlines():
            # Evaluar la línea como una expresión de Python y agregarla a la lista
            filas.append(ast.literal_eval(linea))

    # Crear y devolver un DataFrame a partir de la lista de filas
    return pd.DataFrame(filas)


def expandir_columna_anidada(df_entrada, columna_anidada):
    """
    Expande una columna anidada en un DataFrame.

    Parámetro:
    - df_entrada (pd.DataFrame): DataFrame de entrada con la columna anidada.
    - columna_anidada (str): Nombre de la columna anidada que se va a expandir.

    Retorno:
    pd.DataFrame: Nuevo DataFrame con la columna anidada expandida.
    """
    # Inicializar una lista vacía para almacenar los registros expandidos.
    registros_expandidos = []

    # Iterar sobre cada fila del DataFrame de entrada.
    for indice, fila in df_entrada.iterrows():
        # Extraer los datos base de la fila eliminando la columna anidada y convirtiéndolos a un diccionario.
        datos_base = fila.drop(columna_anidada).to_dict()
        # Obtener la lista de datos anidados de la columna especificada.
        datos_anidados_lista = fila[columna_anidada]

        # Iterar sobre cada conjunto de datos anidados.
        for datos_anidados in datos_anidados_lista:
            # Crear un nuevo registro combinando los datos base y los datos anidados.
            nuevo_registro = {**datos_base, **datos_anidados}
            # Agregar el nuevo registro a la lista de registros_expandidos.
            registros_expandidos.append(nuevo_registro)

    # Crear un nuevo DataFrame a partir de la lista de registros_expandidos y devolverlo.
    df_expandido = pd.DataFrame(registros_expandidos)
    return df_expandido

# Ruta del archivo JSON
ruta_json= "australian_user_reviews.json"
#ruta_json = "..\\Datsets\\australian_user_reviews.json"

# Cargar el archivo JSON en un DataFrame
user_reviews = cargar_json(ruta_json)

# Expandir los diccionarios en varias filas
user_reviews= expandir_columna_anidada(user_reviews, 'reviews')


user_reviews.head()

user_reviews.isnull().sum()

user_reviews.drop_duplicates(inplace=True)
user_reviews.shape

#Eliminamos la columna user_reviews y la columna helpful
user_reviews = user_reviews.drop('user_url', axis=1)
user_reviews = user_reviews.drop('helpful', axis=1)

# Identificar y contar las filas que contienen solo espacios o caracteres en blanco en la columna 'funny'
filas_espacios_funny = user_reviews[user_reviews['funny'].apply(lambda x: x.isspace() or not x)].index

# Imprimir el resultado
print("Número de filas que contienen caracteres en blanco en la columna 'funny':", len(filas_espacios_funny))

# Identificar y contar las filas que contienen solo espacios o caracteres en blanco en la columna 'last_edited'
filas_espacios_last = user_reviews[user_reviews['last_edited'].apply(lambda x: x.isspace() or not x)].index

# Imprimir el resultado
print("Número de filas que contienen caracteres en blanco en la columna 'last_edited':", len(filas_espacios_last))

#Eliminamos la columa funny
user_reviews = user_reviews.drop('funny', axis=1)
#Eliminamos la columa last_edited
user_reviews = user_reviews.drop('last_edited', axis=1)

#Quitamos la palabra Posted 
user_reviews['posted'] = user_reviews['posted'].str.replace('Posted ', '')
#Convertimos a fecha
user_reviews['posted'] = pd.to_datetime(user_reviews['posted'], errors='coerce')
#Extraemos el año y creamos otra columna llamada posted_year
user_reviews['posted_year'] = user_reviews['posted'].dt.year
#Borramos la columna posted
user_reviews=user_reviews.drop('posted',axis=1)

# Ruta del archivo JSON
ruta_json = "australian_users_items.json"

# Cargar el archivo JSON en un DataFrame
user_items = cargar_json(ruta_json)

# Expandir los diccionarios en varias filas
user_items = expandir_columna_anidada(user_items, "items")


user_items.head()

user_items.isnull().sum()

user_items.drop_duplicates(inplace=True)
user_items.shape

#Eliminamos la columa user_url
user_items = user_items.drop('user_url', axis=1)
user_items = user_items.drop('playtime_2weeks', axis=1)
user_items = user_items.drop('steam_id', axis=1)

# Ruta del archivo JSON
ruta_json = "output_steam_games.json"

# Cargar el archivo JSON en un DataFrame
steam_games=pd.read_json(ruta_json, lines=True)

steam_games.head()


steam_games.isnull().sum()

#Borramos las filas en las que todos sus valores sean nulos
steam_games.dropna(how='all', inplace=True)
#Contamos cuantos valores nulos hay por columna
steam_games.isnull().sum()

# Convertimos la columna 'genres' a tipo lista, conservando NaN
steam_games['genres'] = steam_games['genres'].apply(lambda x: ast.literal_eval(x) if isinstance(x, str) else x)
# Expandimos las listas en filas
steam_games_genres = steam_games.explode('genres')
# Obtener la lista de géneros únicos, incluyendo los valores nulos
lista_generos = steam_games_genres['genres'].unique()
steam_games['tags'] = steam_games['tags'].apply(lambda x: ast.literal_eval(x) if isinstance(x, str) else x)
steam_games['genres'] = steam_games.apply(
    lambda row: [tag for tag in row['tags'] if tag in lista_generos] if isinstance(row['tags'], list) else row['genres'],
    axis=1
)

#Lista de desarolladoras
lista_desarroladores =steam_games['developer'].unique()


for index, row in steam_games.iterrows():
    # Verificar si el valor en 'developer' está vacío y si 'publisher' está en la lista de desarrolladores
    if pd.isnull(row['developer']) and row['publisher'] in lista_desarroladores:
        # Rellenar 'developer' con el valor de 'publisher'
        steam_games.at[index, 'developer'] = row['publisher']

# Verificamos filas duplicadas en el dataframe
duplicados = steam_games[steam_games.duplicated(['publisher', 'app_name', 'title', 'url', 
                                                                 'release_date', 'reviews_url', 'price', 'early_access', 
                                                                 'id', 'developer'])]
print(f"Cantidad de filas duplicadas: {len(duplicados)}")
subset_cols = ['publisher', 'app_name', 'title', 'url', 'release_date', 'reviews_url', 'price', 'early_access', 'id', 'developer']
steam_games.drop_duplicates(subset=subset_cols, inplace=True)

steam_games = steam_games.drop('reviews_url', axis=1)
steam_games = steam_games.drop('url', axis=1)
steam_games = steam_games.drop('publisher', axis=1)
steam_games = steam_games.drop('title', axis=1)

#Convertimos la columa release_date a formato fecha
steam_games['release_date']=pd.to_datetime(steam_games['release_date'], errors='coerce', exact=False)
# Extraer el año y crear una nueva columna 'release_year'
steam_games['release_year'] = steam_games['release_date'].dt.year
# Eliminar la columna 'release_date'
steam_games = steam_games.drop(columns=['release_date'])


steam_games['price'] = steam_games['price'].replace('Free To Play', 0)
steam_games['price'] = steam_games['price'].replace('Free to Play', 0)
steam_games['price'] = steam_games['price'].replace('Free', 0)
steam_games['price'] = steam_games['price'].replace('None', 0)
steam_games['price'] = steam_games['price'].replace('Free HITMAN™ Holiday Pack', 0)
steam_games['price'] = steam_games['price'].replace('Free Demo', 0)
steam_games['price'] = steam_games['price'].replace('Play for Free!', 0)
steam_games['price'] = steam_games['price'].replace('Play WARMACHINE: Tactics Demo', 0)
steam_games['price'] = steam_games['price'].replace('Install Now', 0)
steam_games['price'] = steam_games['price'].replace('Install Theme', 0)
steam_games['price'] = steam_games['price'].replace('Third-party', 0)
steam_games['price'] = steam_games['price'].replace('Play Now', 0)
steam_games['price'] = steam_games['price'].replace('Play the Demo', 0)
steam_games['price'] = steam_games['price'].replace('Free to Try', 0)
steam_games['price'] = steam_games['price'].replace('Free to Use', 0)
steam_games['price'] = steam_games['price'].replace('Free Movie', 0)
steam_games['price'] = steam_games['price'].replace('Free Mod', 0)
steam_games['price'] = steam_games['price'].replace('Free Mod', 0)
steam_games['price'] = steam_games['price'].replace('Starting at $499.00', 499)
steam_games['price'] = steam_games['price'].replace('Starting at $449.00', 449)

# Convertir el DataFrame 'steam_games' a CSV --- CSV
steam_games.to_csv("E:\HENRY\PI MLOps\PI MLOps - STEAM\CSV\\steam_games.csv", index=False)

# Convertir el DataFrame 'user_reviews' a CSV
user_reviews.to_csv("E:\HENRY\PI MLOps\PI MLOps - STEAM\CSV\\user_reviews.csv", index=False)

# Convertir el DataFrame 'user_items' a CSV
user_items.to_csv("E:\HENRY\PI MLOps\PI MLOps - STEAM\CSV\\user_items.csv", index=False)


# Convertir el DataFrame 'user_reviews' a CSV
user_reviews.to_csv("E:\HENRY\PI MLOps\PI MLOps - STEAM\CSV\\user_reviews.csv", index=False)

# Convertir el DataFrame 'user_items' a CSV
user_items.to_csv("E:\HENRY\PI MLOps\PI MLOps - STEAM\CSV\\user_items.csv", index=False)
