==> 2. analisis_sentimientos.ipynb <==
import pandas as pd
from nltk.sentiment import SentimentIntensityAnalyzer
import nltk
nltk.download('vader_lexicon')

# Leemos el archivo

user_reviews = pd.read_csv(r"CSV\user_reviews.csv")


def asignar_etiqueta(sentimiento):
    """
    Función para Asignar Etiquetas Numéricas a los Sentimientos

    Esta función recibe un puntaje de sentimiento como entrada y asigna etiquetas numéricas según umbrales predefinidos.

    Parámetros:
    - sentimiento (float): Puntaje de sentimiento obtenido del análisis.

    Retorno:
    - int: Etiqueta numérica asignada, donde:
           - 2 indica sentimiento positivo,
           - 0 indica sentimiento negativo, y
           - 1 indica neutralidad.

    Ejemplo de Uso:
    ```python
    puntaje_sentimiento = analizar_sentimiento(texto)
    etiqueta_asignada = asignar_etiqueta(puntaje_sentimiento)
    print(f"Etiqueta Asignada: {etiqueta_asignada}")
    ```
    """
    if sentimiento >= 0.05:
        return 2  # Sentimiento positivo
    elif sentimiento <= -0.04:
        return 0  # Sentimiento negativo
    else:
        return 1  # Neutral

# Rellenar valores nulos con una cadena vacía
user_reviews['review'].fillna('', inplace=True)

# Inicializar el analizador de sentimientos de NLTK
sid = SentimentIntensityAnalyzer()

# Aplicar el análisis de sentimiento y asignar etiquetas numéricas
user_reviews['sentiment_score'] = user_reviews['review'].apply(lambda x: sid.polarity_scores(x)['compound'])
user_reviews['sentiment_analysis'] = user_reviews['sentiment_score'].apply(asignar_etiqueta)

# Borramos la columa reviews
user_reviews.drop(columns=["review"],inplace=True)

# Convertir el DataFrame 'user_reviews' a CSV
user_reviews.to_csv("E:\HENRY\PI MLOps\PI MLOps - STEAM\CSV\\user_reviews.csv", index=False)
