==> 4.user_for_genre.ipynb <==
import pandas as pd
import ast

steam_games = pd.read_csv(r"CSV\steam_games.csv")
user_items =pd.read_csv(r"CSV\user_items.csv")


# Unir los DataFrames usando las columnas 'id' en steam_games y 'item_id' en user_items
df_UserForGenre = pd.merge(user_items, steam_games, left_on='item_id', right_on='id', how='inner')

# Seleccionar las columnas requeridas
df_UserForGenre = df_UserForGenre[['genres', 'release_year', 'playtime_forever', 'user_id']]

df_UserForGenre.head()

# Verificar si 'genres' contiene NaN antes de aplicar ast.literal_eval
df_UserForGenre['genres'] = df_UserForGenre['genres'].apply(lambda x: ast.literal_eval(x) if pd.notna(x) else [])
# Expandir las listas en columnas
df_UserForGenre = df_UserForGenre.explode('genres').reset_index(drop=True)
df_UserForGenre.head()

# Encontrar el user_id con la mayor cantidad de horas jugadas por género
jugador_mas_horas = df_UserForGenre.groupby(['genres', 'user_id'])['playtime_forever'].sum().reset_index()

# Encontrar el índice del user_id con la mayor cantidad de horas jugadas por género
indice = jugador_mas_horas.groupby('genres')['playtime_forever'].idxmax()

# Filtrar el DataFrame original usando los índices encontrados
jugador_mas_horas = jugador_mas_horas.loc[indice]


jugador_mas_horas.head()

# Concatenamos los DataFrames
df_UserForGenre = pd.merge(df_UserForGenre, jugador_mas_horas[['genres', 'user_id']], on=['genres','user_id'])
df_UserForGenre.head()

df_UserForGenre = df_UserForGenre[df_UserForGenre['playtime_forever']>0]
df_UserForGenre.head()

df_UserForGenre = df_UserForGenre.groupby(['genres', 'release_year','user_id'])['playtime_forever'].sum().reset_index()
df_UserForGenre.head()

#Convertimos a CSV
df_UserForGenre.to_csv(r'E:\HENRY\PI MLOps\PI MLOps - STEAM\Api\UserForGenre.csv', index=False)

def UserForGenre(genre):
    # Filtrar el DataFrame por el género dado
    genero = df_UserForGenre[df_UserForGenre['genres'] == genre]

    # Crear una lista de diccionarios para las horas jugadas por año
    horas_anio = [{'Año': int(year), 'Horas': int(playtime)} for year, playtime in genero[['release_year', 'playtime_forever']].values]

    # Crear el diccionario de salida
    resultado = {
        'Usuario con más horas jugadas para género '+ genre: genero.iloc[0]['user_id'],
        "Horas jugadas": horas_anio
    }

    return resultado

genre_result = UserForGenre('Action')
print(genre_result)
