==> 5.user_recommend.ipynb <==

import pandas as pd

steam_games = pd.read_csv("CSV\steam_games.csv")
user_reviews =pd.read_csv(r"CSV\user_reviews.csv")

# Combinar DataFrames basado en las columnas id y item_id
df_UsersRecommend = pd.merge(user_reviews, steam_games, left_on='item_id', right_on='id', how='inner')
df_UsersRecommend.head()

# Seleccionar las columnas necesarias
selected_columns = ['app_name', 'posted_year', 'recommend', 'sentiment_analysis']
df_UsersRecommend = df_UsersRecommend[selected_columns]
df_UsersRecommend.head()

#Filtramos los comentarios positivos y neutrales
df_UsersRecommend =df_UsersRecommend[df_UsersRecommend['sentiment_analysis']!=0]
df_UsersRecommend.head()

#Agrupamos por app_name , posted_year y sumamos la cantidad de recomendaciones
df_UsersRecommend = df_UsersRecommend.groupby(['app_name', 'posted_year']).agg({'recommend': 'sum'}).reset_index()
df_UsersRecommend = df_UsersRecommend.sort_values('recommend',ascending=False)
df_UsersRecommend.head()


df_UsersRecommend.to_csv(r'Api\UsersRecommend.csv', index=False)

def UsersRecommend(anio):
    """
    Devuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado.
    
    Args:
    - anio (int): El año para el cual se desean obtener las recomendaciones.

    Returns:
    - dict: Un diccionario que contiene el top 3 de juegos recomendados en el formato:
        {"Puesto 1": "Nombre del Juego1", "Puesto 2": "Nombre del Juego2", "Puesto 3": "Nombre del Juego3"}
        En caso de no haber recomendaciones para el año especificado, devuelve un mensaje indicando esto.
    """

    # Filtrar las recomendaciones para el año dado y recomendaciones positivas/neutrales
    recomendaciones = df_UsersRecommend[df_UsersRecommend['posted_year'] == anio]

    # Verificar si hay revisiones para el año dado
    if recomendaciones.empty:
        return f"No hay recomendaciones para el año {anio}"

    # Ordenar en orden descendente por la cantidad de recomendaciones
    recomendaciones = recomendaciones.sort_values('recommend', ascending=False)

    # Crear una única línea de resultado
    resultado = {
        "Puesto 1": recomendaciones.iloc[0]['app_name'],
        "Puesto 2": recomendaciones.iloc[1]['app_name'],
        "Puesto 3": recomendaciones.iloc[2]['app_name']
    }

    return resultado

UsersRecommend(2011)
