==> 6.user_worst_developer.ipynb <==
import pandas as pd

steam_games = pd.read_csv("CSV\steam_games.csv")
user_reviews =pd.read_csv(r"CSV\user_reviews.csv")

# Combinar dataframes en base a las columnas id e item_id
df_UserWorstDeveloper = pd.merge(steam_games, user_reviews, left_on='id', right_on='item_id')

# Seleccionar columnas 
df_UserWorstDeveloper = df_UserWorstDeveloper[['developer', 'recommend', 'sentiment_analysis','posted_year']]


#Filtramos los comentarios negativos
df_UserWorstDeveloper = df_UserWorstDeveloper[df_UserWorstDeveloper['sentiment_analysis'] == 0]


#Agrupamos las recomendaciones por desarrollador y por año
df_UserWorstDeveloper = df_UserWorstDeveloper.groupby(['developer', 'posted_year'])['recommend'].sum().reset_index()
df_UserWorstDeveloper.head()

df_UserWorstDeveloper.to_csv(r'Api\UsersWorstDeveloper.csv', index=False)


def UserWorstDeveloper(anio):
    desarrollador = df_UserWorstDeveloper[df_UserWorstDeveloper['posted_year']== anio]
    if not desarrollador.empty:
        # Encuentra los años con menos recomendaciones
        desarrollador = desarrollador.nsmallest(3, 'recommend')
        resultado = [{"Puesto 1" : desarrollador.iloc[0]['developer']}, 
                  {"Puesto 2" : desarrollador.iloc[1]['developer']},
                  {"Puesto 3" : desarrollador.iloc[2]['developer']}]
        return resultado
    else:
        return f"No se encontraron datos para el año '{anio}'"
    

UserWorstDeveloper(2012)
