==> 7.sentiment_analysis.ipynb <==
import pandas as pd

steam_games = pd.read_csv("CSV\steam_games.csv")
user_reviews =pd.read_csv(r"CSV\user_reviews.csv")

#Concatenamos los dataframes
analisis_sentimientos_df = steam_games[['id', 'developer']].merge(user_reviews[['item_id', 'sentiment_analysis']], left_on='id', right_on='item_id', how='inner')
#Quitamos las columnas id e item_id porque no las vamos a ocupar
analisis_sentimientos_df = analisis_sentimientos_df.drop(columns=['id', 'item_id'])


analisis_sentimientos_df.head()

# Convertir los valores en la columna 'sentiment_analysis' a int
analisis_sentimientos_df['sentiment_analysis'] = analisis_sentimientos_df['sentiment_analysis'].astype(int)
#Cambiamos los valores por positivo , negativo y neutral
analisis_sentimientos_df['sentiment_analysis'] = analisis_sentimientos_df['sentiment_analysis'].replace({0: 'Negativo', 1: 'Neutral', 2: 'Positivo'})


analisis_sentimientos_df.head()

# Agrupar los datos y contar los valores
analisis_sentimientos_df = analisis_sentimientos_df.groupby('developer')['sentiment_analysis'].value_counts()


analisis_sentimientos_df.head()

#Convertimos a CSV
analisis_sentimientos_df.to_csv('Api\sentiment_analysis.csv', index=False)


def sentiment_analysis(developer):
    """
    Retorna un diccionario que muestra la cantidad de reseñas positivas, negativas y neutrales para un desarrollador específico.

    Parameters:
    - developer (str): El nombre del desarrollador para el cual se desea obtener el análisis de sentimientos.

    Returns:
    - dict: Un diccionario con el nombre del desarrollador como clave y una lista con la cantidad total de registros
            de reseñas de usuarios categorizados con un análisis de sentimiento como valor.
            Ejemplo de retorno: {'Valve': {'Negative': 182, 'Neutral': 120, 'Positive': 278}}

            En caso de que no se encuentren datos para el desarrollador especificado, se retorna un mensaje informativo.
            Ejemplo: "No se encontraron datos para el desarrollador 'NombreDelDesarrollador'".
    """
    developer_column = analisis_sentimientos_df["developer"].tolist()

    if developer in developer_column:
        index = developer_column.index(developer)
        sentiment_counts = analisis_sentimientos_df.iloc[index, 1:].to_dict()
        return {developer: sentiment_counts}
    else:
        return f"No se encontraron datos para el desarrollador '{developer}'"
